app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
prometheus-api-client==0.5.3
snowflake-connector-python==3.6.0
python-dotenv==1.0.0
